Refactor D3D12 buffer General code clean up

+++ GraphicsEngine/Source/Core/RHI/Shader.cpp
@@ -86,7 +86,7 @@ int Shader::GetSlotForName(const std::string & name)
 	return -1;
 }
 
-int Shader::GetNameHash()
+uint64 Shader::GetNameHash()
 {
 	if (Hash == 0)
 	{
